dev_requirements.txt
Pillow==2.4.0
Django==1.6.5
django-nose==1.2
django-dotenv==1.2
django-getenv==1.3.1
django-postman==3.1.0
django-facebook==6.0.1
django-allauth==0.18.0
psycopg2==2.5.3
south==1.0

mock==1.0.1
factory_boy==2.3.1
fake-factory==0.4.0
Faker==0.0.4
selenium==2.42.1